Move new decorator to where it belongs and add a test
